Calculate Log Probability of Multiple Prey Remaining States

Computes the log probability of observing specific numbers of prey remaining in a simulation based on a model, given an initial number of prey and other parameters. This function uses simulated data to estimate the log probability of each state using a probability mass function (PMF) and aggregates these probabilities for a list of prey remaining states.

Usage

log_probability_single_prey_initial(
  parameters,
  ns_prey_remaining,
  n_prey_initial,
  model,
  time_max,
  n_replicates = 1000,
  alpha = 1
)

Arguments

parameters

A list or vector containing parameters required by the model function. The specific parameters depend on the model used.

ns_prey_remaining

A numeric vector of integers representing the different states of prey remaining for which the log probability needs to be calculated.

n_prey_initial

An integer specifying the initial number of prey. This is used to define the range of possible values for prey remaining.

model

A function that computes the propensity or rate of change based on the number of prey remaining and the parameters provided. The model should be compatible with the simulation function.

time_max

A numeric value representing the maximum time for the simulation.

n_replicates

An integer specifying the number of simulation runs to perform. Default is 1000.

alpha

A numeric parameter for the Dirichlet process prior. Default is 1.

Value

A numeric value representing the total log probability of observing the specified states of prey remaining. This is the sum of the log probabilities for each value in ns_prey_remaining.

Details

The function first generates simulated data using a given model and parameters, then creates a log PMF function from this data. It calculates the log probability for each specified number of prey remaining by summing the log PMF values.

Examples

# Example usage
parameters <- list(rate = 0.1)  # Example parameter for a constant rate model
ns_prey_remaining <- c(0, 1, 2)  # States of prey remaining
n_prey_initial <- 2
model <- model_constant_rate()  # Using a constant rate model
time_max <- 10

# Compute log probabilities
log_prob <- frbayes:::log_probability_single_prey_initial(
  parameters = parameters,
  ns_prey_remaining = ns_prey_remaining,
  n_prey_initial = n_prey_initial,
  model = model,
  time_max = time_max,
  n_replicates = 1000,
  alpha = 1
)